Extract delivery logic to a module
